randominzed testing

+++ microwave/microwave_generator_readout_V0.py
@@ -205,7 +205,7 @@ def readMicrowave(ser, mode='normal', readline_buffer=500):
 	if master_mode == 'testing':
 		print('Testing inside readMicrowave: str(ser.readline(readline_buffer)) cmd=' + cmd)
 		rand = random.uniform(0, 1)
-		response = '$PWRG:%(val1)s.%(val2)sW' % {"val1": rand*500, "val2": np.rint(rand*10)}
+		response = '$PWRG:%(val1)s.%(val2)sW' % {"val1": rand*500, "val2": int(np.rint(rand*10))}
 	else:
 		response = str(ser.readline(readline_buffer))
 	t_ = re.findall(r':(.+)W', response)
